A few tweaks:
- Added control over saving bad junctions to full pipeline
- Made it clearer that the filter stage requires a junction tab file as input
- Made index numbers 0-based across output files
- Disabled TXT output.  This seems redundant, no one ever uses it and it takes lots of storage space.
- Disabled GFF output.  Not normally required and takes a lot of storage space.  I'll replace this with a mode in the portcullis_convert script later.
- Better description for full pipeline... indicates ML filtering will be applied.
- Fixed a mistake in the threshold option in the filtering tool
- Fixed a mistake in the canonical option in the full and filtering tools.

+++ scripts/bed12.py
@@ -139,7 +139,7 @@ class BedEntry:
 		b.start = int(parts[1])
 		b.end = int(parts[2])
 		b.name = parts[3]
-		b.score = int(parts[4])
+		b.score = float(parts[4])
 		b.strand = parts[5]
 		b.thick_start = int(parts[6])
 		b.thick_end = int(parts[7])
@@ -180,7 +180,7 @@ class BedEntry:
 		b.start = int(parts[6])
 		b.end = int(parts[7]) + 1
 		b.strand = parts[12]
-		b.score = int(parts[14])
+		b.score = float(parts[14])
 		b.thick_start = int(parts[4])
 		b.thick_end = int(parts[5]) + 1
 		b.name = "junc"
